Revert "Improved automatic pre-processing"
This reverts commit a3c6be58ddd1f4fad934379d8f4dd2b74f185407.

+++ propinfer/generator.py
@@ -1,4 +1,4 @@
-from numpy import array, eye, float32
+from numpy import array, eye
 from numpy.random import multivariate_normal
 from pandas import DataFrame, concat
 from sklearn.model_selection import StratifiedShuffleSplit
@@ -36,8 +36,7 @@ class GaussianGenerator(Generator):
         for i in range(1, 5):
             cov[0, i] = cov[i, 0] = 0.5
 
-        data = DataFrame(data=multivariate_normal(mean, cov, size=self.num_samples),
-                         columns=['label', 'f1', 'f2', 'f3', 'f4'], dtype=float32)
+        data = DataFrame(data=multivariate_normal(mean, cov, size=self.num_samples), columns=['label', 'f1', 'f2', 'f3', 'f4'])
         data['label'] = (data['label'] > 0).astype('int32')
 
         return data
@@ -53,8 +52,7 @@ class IndependentPropertyGenerator(Generator):
         for i in range(1, 4):
             cov[0, i] = cov[i, 0] = 0.5
 
-        data = DataFrame(data=multivariate_normal(mean, cov, size=self.num_samples),
-                         columns=['label', 'f1', 'f2', 'f3','f4'], dtype=float32)
+        data = DataFrame(data=multivariate_normal(mean, cov, size=self.num_samples), columns=['label', 'f1', 'f2', 'f3', 'f4'])
         data['label'] = (data['label'] > 0).astype('int32')
 
         return data
